requirements.txt
altgraph==0.17
antlr4-python3-runtime==4.9
APScheduler==3.6.3
apt-clone==0.2.1
apturl==0.5.2
astroid==2.4.2
beautifulsoup4==4.8.2
blinker==1.4
Brlapi==0.7.0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
colorama==0.4.3
command-not-found==0.3
configobj==5.0.6
cryptography==2.8
cupshelpers==1.0
dbus-python==1.2.16
decorator==4.4.2
defer==1.0.6
distro==1.4.0
entrypoints==0.3
grpcio==1.16.1
httplib2==0.14.0
idna==2.8
ifaddr==0.1.6
isort==5.6.4
keyring==18.0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lazy-object-proxy==1.4.3
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==1.1.0
mccabe==0.6.1
nemo-emblems==4.6.0
netifaces==0.10.4
nose==1.3.7
oauthlib==3.1.0
onboard==1.4.1
packaging==20.3
PAM==0.4.2
pexpect==4.6.0
Pillow==7.0.0
protobuf==3.6.1
psutil==5.5.1
pyactiveresource==2.2.1
pycairo==1.16.2
pycrypto==2.6.1
pycups==1.9.73
pycurl==7.43.0.2
PyGObject==3.36.0
PyICU==2.4.2
pyinotify==0.9.6
pyinstaller==4.1
pyinstaller-hooks-contrib==2020.10
PyJWT==1.7.1
pylint==2.6.0
pymacaroons==0.13.0
PyNaCl==1.3.0
pyparsing==2.4.6
pyparted==3.11.2
pyRFC3339==1.1
python-apt==2.0.0+ubuntu0.20.4.1
python-debian===0.1.36ubuntu1
python-telegram-bot==13.1
python-xapp==2.0.1
python-xlib==0.23
pytz==2019.3
pyxdg==0.26
PyYAML==5.3.1
reportlab==3.5.34
requests==2.22.0
requests-unixsocket==0.2.0
screen-resolution-extra==0.0.0
SecretStorage==2.3.1
setproctitle==1.1.10
ShopifyAPI==8.1.0
simplejson==3.16.0
six==1.14.0
soupsieve==1.9.5
systemd-python==234
tinycss2==1.0.2
toml==0.10.2
tornado==6.1
tzlocal==2.1
ubuntu-advantage-tools==20.3
ubuntu-drivers-common==0.0.0
ufw==0.36
Unidecode==1.1.1
urllib3==1.25.8
wadllib==1.3.3
webencodings==0.5.1
wrapt==1.12.1
xkit==0.0.0
zeroconf==0.24.4
